+++ src/libopenrave-core/xmlreaders-core.cpp
@@ -1404,6 +1404,12 @@ public:
         case PE_Ignore: return PE_Ignore;
         }
 
+        if( xmlname == "actuator" ) {
+            _processingtag = "";
+            _pcurreader.reset(new xmlreaders::ElectricMotorActuatorInfoReader(ElectricMotorActuatorInfoPtr(), atts));
+            return PE_Support;
+        }
+
         static boost::array<string, 23> tags = { { "body", "offsetfrom", "weight", "lostop", "histop", "limits", "limitsrad", "limitsdeg", "maxvel", "maxveldeg", "hardmaxvel", "maxaccel", "maxacceldeg", "maxtorque", "maxforce", "resolution", "anchor", "axis", "axis1", "axis2", "axis3", "mode", "initial" }};
         if( find(tags.begin(),tags.end(),xmlname) != tags.end() ) {
             _processingtag = xmlname;
@@ -1419,6 +1425,10 @@ public:
 
         if( !!_pcurreader ) {
             if( _pcurreader->endElement(xmlname) ) {
+                xmlreaders::ElectricMotorActuatorInfoReaderPtr actuatorreader = boost::dynamic_pointer_cast<xmlreaders::ElectricMotorActuatorInfoReader>(_pcurreader);
+                if( !!actuatorreader ) {
+                    _pjoint->_info._infoElectricMotor = actuatorreader->GetActuatorInfo();
+                }
                 _pcurreader.reset();
             }
         }